+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortToolMakerTemplate.java
@@ -1,0 +1,13 @@
+package ai.mcpdirect.backend.dao.entity.aitool;
+
+public class AIPortToolMakerTemplate {
+    public long id;
+    public long created;
+    public int status;
+    public int type;
+    public String name;
+    public String tags;
+    public long agentId;
+    public long lastUpdated;
+    public long userId;
+}